Move Test to eidoscommon
